feat: add RightBody component
