+++ src/App.tsx
@@ -1,22 +1,12 @@
-import React, { useCallback } from "react";
+import React from "react";
 
-import Header from "@/components/Header/Header";
-import SearchPage from "@/components/SearchPage/SearchPage";
-import { Wrapper, SearchPageWrapper, AddButton } from "./App.styled";
+import RightBody from "@/components/RightBody/RightBody";
+import { Wrapper } from "./App.styled";
 
 function App() {
-  const handleAddButtonClick = useCallback(() => {
-    console.log("click");
-  }, []);
-
   return (
     <Wrapper>
-      <Header />
-      <SearchPageWrapper>
-        <img src="kangaroo.png" />
-        <SearchPage />
-        <AddButton src="addbtn.png" onClick={handleAddButtonClick} />
-      </SearchPageWrapper>
+      <RightBody />
     </Wrapper>
   );
 }
