+++ src/components/RightBody/RightBody.styled.ts
@@ -1,0 +1,16 @@
+import styled from "styled-components";
+
+export const SearchPageWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: space-around;
+  align-items: center;
+  width: 100%;
+  height: 300px;
+`;
+
+export const AddButton = styled.img`
+  :hover {
+    cursor: pointer;
+  }
+`;
